backend/requirements.txt
schwab-py>=1.4.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
python-dotenv>=1.0.0
numpy>=1.26.0
scipy>=1.13.0
apscheduler>=3.10.0
pytz>=2024.1
slowapi>=0.1.9
